refactor(envoy): move admin client to plugin envoy
