+++ src/cactus_runner/app/action.py
@@ -12,9 +12,6 @@ from envoy_schema.server.schema.sep2.types import DeviceCategory
 # from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-#
-from cactus_runner.app.envoy_admin_client import EnvoyAdminClient
-
 # from cactus_runner.app.envoy_common import (
 #    count_all_site_controls_with_cancelled,
 #    get_active_site,
@@ -34,6 +31,7 @@ from cactus_runner.models import (
 )
 from cactus_runner.plugin import dtos
 from cactus_runner.plugin.backends.common import RunnerBackend, get_site_control_groups_ordered
+from cactus_runner.plugin.backends.envoy import EnvoyAdminClient
 from cactus_runner.plugin.backends.hookspec import create_backend
 
 logger = logging.getLogger(__name__)
