+++ src/cactus_runner/app/event.py
@@ -9,9 +9,9 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from cactus_runner.app import evaluator
 from cactus_runner.app.action import apply_actions
 from cactus_runner.app.check import all_checks_passing
-from cactus_runner.app.envoy_admin_client import EnvoyAdminClient
 from cactus_runner.app.uri import MountedProxyPathParts, does_endpoint_match
 from cactus_runner.models import ActiveTestProcedure, Listener, RunnerState
+from cactus_runner.plugin.backends.envoy import EnvoyAdminClient
 
 logger = logging.getLogger(__name__)
