+++ src/cactus_runner/plugin/backends/envoy/__init__.py
@@ -1,3 +1,4 @@
+from cactus_runner.plugin.backends.envoy.admin_client import EnvoyAdminClient
 from cactus_runner.plugin.backends.envoy.backend import EnvoyBackend
 
-__all__ = ["EnvoyBackend"]
+__all__ = ["EnvoyBackend", "EnvoyAdminClient"]
