+++ tests/unit/app/test_envoy.py
@@ -22,7 +22,7 @@ from envoy_schema.admin.schema.uri import (
     SiteControlUri,
 )
 
-from cactus_runner.app.envoy_admin_client import (
+from cactus_runner.plugin.backends.envoy.admin_client import (
     EnvoyAdminClient,
     EnvoyAdminClientAuthParams,
 )
